[TEMPHACK] ci: .shippable.yml: remove some build steps to upload cache
Building with an empty cache currently takes more than one hour which
causes a timeout (Open Source project are allowed a maximum of 1h build
time). As a result, the first build cannot succeed and since the cache
is only saved on success by default, the next build is not faster and
fails as well.

As a workaround,  remove some build steps to allow for a faster build
and therefore populate the cache for the next full builds. Forcing a
build of this branch should hopefully re-create the cache if needed.

Use until [1] is addressed.

Link: [1] https://github.com/Shippable/support/issues 3838
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>